# [controls] fix performance issue in {AppThemeBinding}