@hartez 2023-06-01
/backport to net7.0